centralize basic backup test, add teamschats (#5133)
adds teamchats to the integration backup tests. Moves the common basic
test func into a shared package and calls that test with exchange,
groups, and teamchats.
