+++ src/internal/operations/test/m365/groups/groups_test.go
@@ -4,21 +4,13 @@ import (
 	"context"
 	"testing"
 
-	"github.com/alcionai/clues"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/alcionai/corso/src/internal/events"
-	evmock "github.com/alcionai/corso/src/internal/events/mock"
 	"github.com/alcionai/corso/src/internal/m365/collection/drive"
 	. "github.com/alcionai/corso/src/internal/operations/test/m365"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/internal/tester/tconfig"
-	"github.com/alcionai/corso/src/internal/version"
-	deeTD "github.com/alcionai/corso/src/pkg/backup/details/testdata"
 	"github.com/alcionai/corso/src/pkg/control"
-	"github.com/alcionai/corso/src/pkg/count"
 	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/selectors"
 	selTD "github.com/alcionai/corso/src/pkg/selectors/testdata"
@@ -190,124 +182,14 @@ func runGroupsIncrementalBackupTests(
 		true)
 }
 
-func (suite *GroupsBackupIntgSuite) TestBackup_Run_groupsBasic() {
-	t := suite.T()
-
-	ctx, flush := tester.NewContext(t)
-	defer flush()
-
-	var (
-		mb      = evmock.NewBus()
-		counter = count.New()
-		sel     = selectors.NewGroupsBackup([]string{suite.its.Group.ID})
-		opts    = control.DefaultOptions()
-		whatSet = deeTD.CategoryFromRepoRef
-	)
-
+func (suite *GroupsBackupIntgSuite) TestBackup_Run_basicBackup() {
+	sel := selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 	sel.Include(
 		selTD.GroupsBackupLibraryFolderScope(sel),
 		selTD.GroupsBackupChannelScope(sel),
 		selTD.GroupsBackupConversationScope(sel))
 
-	bo, bod := PrepNewTestBackupOp(t, ctx, mb, sel.Selector, opts, version.Backup, counter)
-	defer bod.Close(t, ctx)
-
-	reasons, err := bod.Sel.Reasons(bod.Acct.ID(), false)
-	require.NoError(t, err, clues.ToCore(err))
-
-	RunAndCheckBackup(t, ctx, &bo, mb, false)
-
-	for _, reason := range reasons {
-		CheckBackupIsInManifests(
-			t,
-			ctx,
-			bod.KW,
-			bod.SW,
-			&bo,
-			bod.Sel,
-			bod.Sel.ID(),
-			reason.Category())
-	}
-
-	_, expectDeets := deeTD.GetDeetsInBackup(
-		t,
-		ctx,
-		bo.Results.BackupID,
-		bod.Acct.ID(),
-		bod.Sel.ID(),
-		path.GroupsService,
-		whatSet,
-		bod.KMS,
-		bod.SSS)
-	deeTD.CheckBackupDetails(
-		t,
-		ctx,
-		bo.Results.BackupID,
-		whatSet,
-		bod.KMS,
-		bod.SSS,
-		expectDeets,
-		false)
-
-	// Basic, happy path incremental test.  No changes are dictated or expected.
-	// This only tests that an incremental backup is runnable at all, and that it
-	// produces fewer results than the last backup.
-	//
-	// Incremental testing for conversations is limited because of API restrictions.
-	// Since graph doesn't provide us a way to programmatically delete conversations,
-	// or create new conversations without a delegated token, we can't do incremental
-	// testing with newly added items.
-	incMB := evmock.NewBus()
-	incBO := NewTestBackupOp(
-		t,
-		ctx,
-		bod,
-		incMB,
-		opts,
-		count.New())
-
-	RunAndCheckBackup(t, ctx, &incBO, incMB, true)
-
-	for _, reason := range reasons {
-		CheckBackupIsInManifests(
-			t,
-			ctx,
-			bod.KW,
-			bod.SW,
-			&incBO,
-			bod.Sel,
-			bod.Sel.ID(),
-			reason.Category())
-	}
-
-	_, expectDeets = deeTD.GetDeetsInBackup(
-		t,
-		ctx,
-		incBO.Results.BackupID,
-		bod.Acct.ID(),
-		bod.Sel.ID(),
-		bod.Sel.PathService(),
-		whatSet,
-		bod.KMS,
-		bod.SSS)
-	deeTD.CheckBackupDetails(
-		t,
-		ctx,
-		incBO.Results.BackupID,
-		whatSet,
-		bod.KMS,
-		bod.SSS,
-		expectDeets,
-		false)
-
-	assert.NotZero(
-		t,
-		incBO.Results.Counts[string(count.PersistedCachedFiles)],
-		"cached items")
-	assert.Greater(t, bo.Results.ItemsWritten, incBO.Results.ItemsWritten, "incremental items written")
-	assert.Greater(t, bo.Results.BytesRead, incBO.Results.BytesRead, "incremental bytes read")
-	assert.Greater(t, bo.Results.BytesUploaded, incBO.Results.BytesUploaded, "incremental bytes uploaded")
-	assert.Equal(t, 1, incMB.TimesCalled[events.BackupEnd], "incremental backup-end events")
+	RunBasicBackupTest(suite, sel.Selector)
 }
 
 type GroupsBackupNightlyIntgSuite struct {
